apply adjustments
The largest change is the use of numpy arrays in place of dictionarys.
Magic numbers have been made into global constants that can be adjusted in the script.
With STATUS MESSAGES and WRI%TE FILES set to true, the only added functionality is the ability to return the data as an array, so that other scripts receive the data directly, rather than reading it in.